+++ R/wfm.R
@@ -938,25 +938,135 @@ Filter.character <- function(x, min = 1, max = Inf, count.apostrophe = TRUE,
 #' 
 #' \code{as.wfm} - Attempts to coerce a matrix to a \code{\link[qdap]{wfm}}.
 #' 
-#' @param matrix.object A matrix object with words for row names and integer 
-#' values.
+#' @param x An object with words for row names and integer values.
+#' @param x Other arguments passed to \code{as.wfm} methods.
 #' @rdname Word_Frequency_Matrix
 #' @export
 #' @return \code{as.wfm} - Returns a matrix of the class "wfm".
-as.wfm <- function(matrix.object) {
+as.wfm <- function(x, ...){
+    
+    x
 
-    if(!all(is.Integer(matrix.object))){
-        stop("matrix.object must contain only integer values")
+    UseMethod("as.wfm")
+}  
+
+#' \code{as.wfm.matrix} - \code{matrix} method for \code{as.wfm} used to 
+#' convert matrices to a \code{wfm}.
+#' @rdname wfm
+#' @export
+#' @method as.wfm matrix
+as.wfm.matrix <- function(x) {
+
+    if(!all(is.Integer(x))){
+        stop("x must contain only integer values")
     }
-    if (any(class(matrix.object) %in% c("TermDocumentMatrix", "DocumentTermMatrix"))) {
-        return(tm2qdap(matrix.object))
-    }
-    if (!is.matrix(matrix.object)) {
-        warning("Not a matrix.object; may not convert correctly",
-            immediate. = TRUE)
-        matrix.object <- as.matrix(matrix.object)
-    } 
-    class(matrix.object) <- c("wfm", "true.matrix", class(matrix.object))
-    matrix.object    
+
+    class(x) <- c("wfm", "true.matrix", class(x))
+    x    
 }
+
+
+#' \code{as.wfm.default} - Default method for \code{as.wfm} used to 
+#' convert matrices to a \code{wfm}.
+#' @rdname wfm
+#' @export
+#' @method as.wfm default 
+as.wfm.default <- function(x) {
+
+    if(!all(is.Integer(x))){
+        stop("x must contain only integer values")
+    }
+    warning("Not a matrix.object; may not convert correctly", immediate. = TRUE)
+    x <- as.matrix(x)
+    
+    class(x) <- c("wfm", "true.matrix", class(x))
+    x    
+}
+
+#' \code{as.wfm.TermDocumentMatrix} - \code{TermDocumentMatrix} method for 
+#' \code{as.wfm} used to a \code{TermDocumentMatrix} to a \code{wfm}.
+#' @rdname wfm
+#' @export
+#' @method as.wfm TermDocumentMatrix 
+as.wfm.TermDocumentMatrix <- function(x) {
+  
+    tm2qdap(x)
+ 
+}
+
+
+#' \code{as.wfm.DocumentTermMatrix} - \code{DocumentTermMatrix} method for 
+#' \code{as.wfm} used to a \code{DocumentTermMatrix} to a \code{wfm}.
+#' @rdname wfm
+#' @export
+#' @method as.wfm DocumentTermMatrix 
+as.wfm.DocumentTermMatrix <- function(x) {
+  
+    tm2qdap(x)
+ 
+}
+
+#' \code{as.wfm.data.frame} - data.frame method for \code{as.wfm} used to 
+#' convert matrices to a \code{wfm}.
+#' @rdname wfm
+#' @export
+#' @method as.wfm data.frame 
+as.wfm.data.frame <- function(x) {
+
+    if(!all(is.Integer(x))){
+        stop("x must contain only integer values")
+    }
+
+    x <- as.matrix(x)
+    
+    class(x) <- c("wfm", "true.matrix", class(x))
+    x    
+}
+
+#' \code{as.wfm.wfdf} - wfdf method for \code{as.wfm} used to 
+#' convert matrices to a \code{wfm}.
+#' @rdname wfm
+#' @export
+#' @method as.wfm wfdf 
+as.wfm.wfdf <- function(x) {
+
+    wfm(x)
+
+}
+
+## tm Package Compatibility Tools: Apply to or Convert to/from Term Document 
+## Matrix or Document Term Matrix
+## 
+## \code{tm2qdap} - Convert the \pkg{tm} package's 
+## \code{\link[tm]{TermDocumentMatrix}}/\code{\link[tm]{DocumentTermMatrix}} to
+## \code{\link[qdap]{wfm}}.
+## 
+## @param x A \code{\link[tm]{TermDocumentMatrix}}/\code{\link[tm]{DocumentTermMatrix}}.
+## @return \code{tm2qdap} - Returns a \code{\link[qdap]{wfm}} object or 
+## \code{weight} object.
+## @rdname tdm
+## INTERNAL HELPER FUNCTION TO CONVERT "DocumentTermMatrix", "TermDocumentMatrix"
+## TO "wfm"
+tm2qdap <- function(x) {
+
+    opts <- c("DocumentTermMatrix", "TermDocumentMatrix")
+    cls <- opts[opts %in% class(x)]
+
+    if (cls == "DocumentTermMatrix") {
+        x <- t(x)
+    }
+    
+    y <- as.matrix(data.frame(as.matrix(x), check.names = FALSE))
+    
+    if(!any(attributes(x)[["Weighting"]] %in% "tf")){
+        class(y) <- c("weighted_wfm", class(y))
+    } else {
+        class(y) <- c("wfm", "true.matrix", class(y))
+    }
+
+    y
+
+}
+
+